+++ CHANGELOG.md
@@ -10,6 +10,7 @@ backwards compatibility.
 * **Breaking** Bump STM32H7 HAL to version 0.14.
 * Improve style and documentation of examples.
 * Introduce RTIC example of audio processing.
+* Introduce example of SD card interaction.
 
 ## 0.7.3